requirements.txt
# Core dependencies
python-telegram-bot==20.3
aiocron>=1.8
openpyxl==3.1.2
pandas==2.2.0  # Для генерации Excel-отчётов
pytz==2023.3
python-dotenv==1.1.0
httpx==0.24.1
filelock==3.16.1  # Для блокировок при работе с файлами
# matplotlib==3.7.1; python_version < '3.12'  # Версия без жестких зависимостей
fast_bitrix24
apscheduler

# Optional for colored logs (Windows)
colorlog==6.7.0

# Watchdog for hot-reloading
watchdog==3.0.0

# Ensure pip version (optional)
pip>=25.0.1
requests>=2.25.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
yadisk>=3.0.0  # Для загрузки бекапов на Яндекс.Диск